Add deadzone
